[time.clock.file] tests now pass, consolidate testing code
