+++ tests/std/test.lst
@@ -235,7 +235,6 @@ tests\P0355R7_calendars_and_time_zones_dates_literals
 tests\P0355R7_calendars_and_time_zones_hms
 tests\P0355R7_calendars_and_time_zones_io
 tests\P0355R7_calendars_and_time_zones_time_point_and_durations
-tests\P0355R7_clocks
 tests\P0356R5_bind_front
 tests\P0357R3_supporting_incomplete_types_in_reference_wrapper
 tests\P0408R7_efficient_access_to_stringbuf_buffer
