--- tests/std/tests/P0355R7_clocks/env.lst
@@ -1,4 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-RUNALL_INCLUDE ..\usual_latest_matrix.lst
